requirements.txt
apscheduler
pyyaml
soco
cherrypy